HTBHF-1863 refactor handle path request
