+++ src/web/routes/application/middleware/handle-path-request/handle-path-request.js
@@ -1,13 +1,9 @@
-const { CHECK_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../common/constants')
+const { CONFIRM_URL } = require('../../common/constants')
 const { stateMachine, actions, states } = require('../../common/state-machine')
+const { isPathInSequence, stepNotNavigable } = require('./predicates')
+const { getPathsInSequence } = require('./selectors')
 
 const { IS_PATH_ALLOWED, GET_NEXT_ALLOWED_PATH } = actions
-
-const getPathsInSequence = (steps) => [...steps.map(step => step.path), CHECK_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL]
-
-const stepNotNavigable = (step, req) => step && typeof step.isNavigable === 'function' && !step.isNavigable(req.session)
-
-const isPathInSequence = (path, sequence) => sequence.includes(path)
 
 const middleware = (config, pathsInSequence, step) => (req, res, next) => {
   // Destroy the session on navigating away from CONFIRM_URL
