+++ src/web/routes/application/middleware/handle-path-request/predicates.js
@@ -1,0 +1,8 @@
+const stepNotNavigable = (step, req) => step && typeof step.isNavigable === 'function' && !step.isNavigable(req.session)
+
+const isPathInSequence = (path, sequence) => sequence.includes(path)
+
+module.exports = {
+  stepNotNavigable,
+  isPathInSequence
+}
